+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 - pastes are now public by default
 - latest public pastes are now shown on the main page
 - `YAP_PASTE_RATE_LIMIT` now controls how often new pastes can be created from the same IP
+- `YAP_NUM_REVERSE_PROXIES` now controls how many reverse proxies are in front of YAP to properly extract paste
+  author IP address from requests
 
 # 0.2.0
 - banner is now a link